requirements.txt
# Core dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.21.0
pillow>=9.0.0

# Audio and speech
speechrecognition>=3.10.0
pyaudio>=0.2.11
gtts>=2.3.0
pygame>=2.5.0
pyttsx3>=2.90

# System control
pyautogui>=0.9.54
pycaw>=20230407
comtypes>=1.1.14
pywin32>=306

# AI and ML
google-generativeai>=0.3.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0

# Jarvis advanced features (optional)
openwakeword>=0.6.0
faster-whisper>=1.0.0
edge-tts>=7.0.0
onnxruntime>=1.15.0
psutil>=5.9.0
soundfile>=0.12.0

# Utilities
requests>=2.31.0
httpx>=0.24.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0

# Missing dependencies found in imports
face-recognition>=1.3.0
sentence-transformers>=2.2.0
librosa>=0.10.0
scikit-learn>=1.3.0
PyJWT>=2.8.0
websockets>=11.0.0
networkx>=3.1
ultralytics>=8.0.0
croniter>=1.4.0
pytz>=2023.3

# Enhanced Voice Synthesis
azure-cognitiveservices-speech>=1.34.0
sounddevice>=0.4.6
webrtcvad>=2.0.10
scipy>=1.11.0

# Smart Home Integration
aiohttp>=3.9.0
pyyaml>=6.0.1
cryptography>=41.0.7
zeroconf>=0.131.0

# Advanced Data Visualization
matplotlib>=3.7.2
plotly>=5.17.0
dash>=2.14.1
seaborn>=0.12.2
pandas>=2.1.3
dash-bootstrap-components>=1.5.0
kaleido>=0.2.1

# OpenGL and Graphics
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
glfw>=2.6.2
moderngl>=5.8.2

# GPU Support (optional)
GPUtil>=1.4.0
pydub>=0.25.1
aiofiles>=23.2.0
asyncio-mqtt>=0.16.1

# Augmented Reality Interface
opencv-contrib-python>=4.8.0  # Extended OpenCV with ArUco markers
glfw>=2.6.2                   # GLFW for OpenGL window management
moderngl>=5.8.2               # Modern OpenGL wrapper
pyrr>=0.10.3                  # 3D math library for OpenGL
trimesh>=4.0.5                # 3D geometry processing
open3d>=0.18.0                # 3D data processing
pyopengl-accelerate>=3.1.7    # OpenGL acceleration
spatial-geometry>=0.2.0       # Spatial geometry utilities
scipy>=1.11.0                 # Scientific computing for SLAM
scikit-image>=0.21.0          # Image processing for AR
imutils>=0.5.4                # Computer vision utilities

# Security & Privacy System
cryptography>=41.0.7          # Comprehensive cryptography library
PyJWT>=2.8.0                  # JSON Web Tokens for secure sessions
qrcode>=7.4.2                 # QR code generation for MFA
pyotp>=2.9.0                  # Time-based one-time passwords
face-recognition>=1.3.0       # Face recognition for biometric auth
bcrypt>=4.1.2                 # Password hashing
keyring>=24.3.0               # Secure credential storage
pycryptodome>=3.19.0          # Additional cryptographic functions
python-jose>=3.3.0           # JSON Object Signing and Encryption
passlib>=1.7.4               # Password hashing utilities